requirements.txt
# Backend & Pipeline
fastapi==0.129.0
uvicorn==0.41.0
paho-mqtt==2.1.0
SQLAlchemy==2.0.46
psycopg2-binary==2.9.11

# Data Science & AI
numpy==2.4.2
pandas==3.0.1
scikit-learn==1.8.0
tensorflow==2.20.0
keras==3.13.2
scipy==1.17.0

# Visualization & Utialities
matplotlib==3.10.8
requests==2.32.5
pydantic==2.12.5
rich==14.3.2
dotenv